주사위 굴리기, 게임방 로직 기초 구현
1. 주사위 굴리기
- 클라이언트 측 본인 차례일 경우, 중앙에 버튼이 출력,

- 버튼을 누르면 서버에 신호를 보내고 랜덤으로 주사위 값을 보내옵니다.

2. 게임방 로직 기초
- 상태값을 이용해 로직을 제어하여 게임에 필요한 코드들을 실행하도록 구현,

->주사위 값 송신까지만 구현,
해당 주사위 값으로 클라이언트 갱신 및 서버 맵 데이터 갱신은 미 처리,

+++ BlueMarble_Client/RenderManager.cpp
@@ -137,7 +137,7 @@ void RenderManager::DrawBoardMap()
         LineTo(memDC, RIGHT_BOTTOM_PRINT_POINT.x, LEFT_TOP_PRINT_POINT.y + (tileHeight * (i+1)));
     }
     
-    for (int i = 0; i < board.name.size(); i++)
+    for (int i = 0; i < (int)board.name.size(); i++)
     {
         DrawText(memDC, board.name[i].c_str(), -1, &rectVector[i], DT_NOCLIP | DT_CENTER);
     }
